% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hamming_dist}
\alias{hamming_dist}
\title{Calculate Hamming distance between two binary strings}
\usage{
hamming_dist(g1, g2)
}
\arguments{
\item{g1}{Character string, binary (e.g. "010").}

\item{g2}{Character string, binary (e.g. "000").}
}
\value{
Integer number of differing bits between `g1` and `g2`.
}
\description{
Calculate Hamming distance between two binary strings
}
\examples{
hamming_dist("000", "010")
}
