% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{interaction_transitions}
\alias{interaction_transitions}
\title{Calculate the number of transitions that change community diversity}
\usage{
interaction_transitions(dataset)
}
\arguments{
\item{dataset}{Adjacency matrix}
}
\value{
A list with three named elements: 
- `total`: Number of transitions in the dataset
- `increasing`: Of which the number of species increases
- `decreasing`: Of which the number of species decreases
}
\description{
Different ecological states may have different numbers of species in them. 
This function sums the number of mutations in the system that cause gain or loss of community diversity.
}
\examples{
dataset <- data.frame(ID = c("(00)", "(01)", "(11)"), fitness = c(0.1, 0.2, 0.3)) \%>\% 
  fitnesses_to_adjacency()
interaction_transitions(dataset)
}
