% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_basin}
\alias{list_basin}
\title{List all states within a basin of accessibility}
\usage{
list_basin(dataset, root)
}
\arguments{
\item{dataset}{Adjacency matrix of transitions.}

\item{root}{Character string; label of the root node.}
}
\value{
Character vector of labels in the same basin as the root.
}
\description{
Traverses graph to find all nodes reachable from a given root.
}
\examples{
dataset <- data.frame(ID = c("(00)", "(01)", "(11)"), fitness = c(0.1, 0.2, 0.3)) \%>\% 
  fitnesses_to_adjacency()
list_basin(dataset, "(11)")
}
