% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_extrema}
\alias{list_extrema}
\title{Identify peaks and valleys in fitness landscape}
\usage{
list_extrema(dataset, input_labels)
}
\arguments{
\item{dataset}{Adjacency matrix of transitions.}

\item{input_labels}{Vector of corresponding node labels.}
}
\value{
A list with two named elements:
- `peaks`: labels with no outgoing edges.
- `valleys`: labels with no incoming edges.
}
\description{
Returns lists of local maxima (peaks) and minima (valleys) based on an adjacency matrix.
}
\examples{
dataset <- data.frame(ID = c("(00)", "(01)", "(11)"), fitness = c(0.1, 0.2, 0.3)) \%>\% 
  fitnesses_to_adjacency()
list_extrema(dataset, colnames(dataset))
}
