% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_shortest_paths}
\alias{list_shortest_paths}
\title{Find shortest paths between two states}
\usage{
list_shortest_paths(dataset, source, target)
}
\arguments{
\item{dataset}{Adjacency matrix.}

\item{source}{Character string giving starting node label.}

\item{target}{Character string giving target node label.}
}
\value{
A list of shortest paths (as strings with arrows between labels), or
`"No possible paths"` if none exist.
}
\description{
Computes all shortest directed paths between two nodes in the transition network.
}
\examples{
dataset <- data.frame(ID = c("(00)", "(01)", "(11)"), fitness = c(0.1, 0.2, 0.3)) \%>\% 
  fitnesses_to_adjacency()
list_shortest_paths(dataset, "(00)", "(11)")
}
