% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{min_distance}
\alias{min_distance}
\title{Minimum number of mutations between two states}
\usage{
min_distance(source, targets)
}
\arguments{
\item{source}{Character string, binary source genotype (e.g. "000").}

\item{targets}{Character string of one or more target genotypes separated by newlines.}
}
\value{
Integer representing the minimal number of mutational steps.
}
\description{
Calculates the minimum number of mutations required to connect a source genotype
to one or more target genotypes, via the shortest path in Hamming space.
}
\examples{
min_distance("000", "010")       # 1
min_distance("000", "010\n111")  # >1
}
