% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_fitness_table}
\alias{parse_fitness_table}
\title{Parse an input fitness table}
\usage{
parse_fitness_table(input_data)
}
\arguments{
\item{input_data}{File path to the input CSV with columns for species and a final column `fitness`.}
}
\value{
A data frame with columns:
\describe{
  \item{fitness}{Numeric fitness values.}
  \item{ID}{Binary string label of concatenated genotypes, e.g. "(00,11)".}
}
}
\description{
Reads and cleans a CSV file of fitness values and species genotype combinations.
Concatenates all genotype columns into a single binary ID column.
}
\details{
Input columns: ("species1 = "00", species2 = "00",..., speciesN, fitness = <dbl>)
Column names for species are unimportant, except for the final 'fitness' column.
Concatenates all provided (potentially variable numbers of) species into one ID label e.g. "(sp1,sp2,...)"
}
\examples{
\dontrun{
df <- parse_fitness_table("data/fitness.csv")
head(df)
}
}
