% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_bitwidth}
\alias{set_bitwidth}
\title{Determine the bit width of a dataset}
\usage{
set_bitwidth(input_column)
}
\arguments{
\item{input_column}{A character vector of binary labels from the dataset (e.g. from an adjacency matrix).}
}
\value{
An integer representing the bit width (number of bits per locus).
}
\description{
Determines the bit width (number of loci) from a dataset’s labels.
Labels must be in binary format, e.g. "(00,00)", and genome size is assumed homogeneous.
}
\examples{
labels <- c("(00,00)", "(11,11)")
set_bitwidth(labels)
}
