% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_labels}
\alias{sort_labels}
\title{Sort binary labels by genetic distance}
\usage{
sort_labels(input_labels)
}
\arguments{
\item{input_labels}{Character vector or named list of binary labels.}
}
\value{
A character vector of labels sorted by increasing genetic distance.
}
\description{
Sorts a set of labels by increasing Hamming distance from the all-zero state.
}
\examples{
labels <- c("(00)", "(01)", "(11)")
sort_labels(labels)
}
